+++ light_util.py
@@ -17,54 +17,7 @@ import time
 import random as rnd
 from ulab import numpy as np
 import rp2
-
-
-@rp2.asm_pio()
-def pulsewidth():
-    wrap_target()
-    wait(1, pin, 0)                       # 0
-    set(x, 0)                             # 1
-    jmp(x_dec, "3")                       # 2
-    label("3")
-    jmp(x_dec, "4")                       # 3
-    label("4")
-    jmp(pin, "3")                         # 4
-    mov(isr, x)                           # 5
-    push(isr, block)                      # 6
-    irq( 0)                               # 7
-    wrap()
-    
-    
-@rp2.asm_pio()
-def pulsewidth2():
-    wrap_target()
-    wait(1, pin, 1)                       # 0
-    set(y, 1)                             # 1
-    jmp(y_dec, "3")                       # 2
-    label("3")
-    jmp(y_dec, "4")                       # 3
-    label("4")
-    jmp(pin, "3")                         # 4
-    mov(isr, y)                           # 5
-    push(isr, block)                      # 6
-    irq( 1)                               # 7
-    wrap()
-
-
-def perc_to_duty(percent):
-    duty = int((percent / 100.0) * 65535.0)
-    return duty
-
-
-def sleep_ms(ms):
-    time.sleep(ms / 1000)
-    return
-
-
-def servo_duty_cycle(pulse_ms, frequency=50):
-    period_ms = 1.0 / frequency * 1000.0
-    duty_cycle = int(pulse_ms / (period_ms / 65535.0))
-    return duty_cycle
+from pwm_util import perc_to_duty
 
 
 def double_backfire(led):
@@ -108,12 +61,6 @@ def single_backfire(led):
     # Return to zero
     led.duty_u16(perc_to_duty(0))
     return
-
-
-def throttle_2_perc(throttle_pwm):
-    dec = (throttle_pwm - 3042) / 895
-    perc = int(dec * 100)
-    return perc
 
 
 def tailights_on(left):
